final/requirements.txt
numpy==1.26.4
opencv-python==4.6.0.66
tqdm==4.66.4
Pillow==10.3.0
matplotlib==3.9.0